Completely working version with login/register and welcome ${user} with username/email and password.
This is the version 4 of reviewit. In this version i fetched the server with email password and username to create account with both email and username. When you are logged in you can post something ( not available feed yet but you can post in different categories and the db gets it). You have a message that sais Welcome ! ${user} and a logout button !